# api/openapi.yaml
openapi: 3.0.2
info:
  title: universal-password-based-authentication-server
  description: Universal password based authentication server.
  contact:
    name: Kandakov Danil
  version: 1.0.0
servers:
- url: http://localhost
  description: For testing, I need it, sorry
...
paths:
...
  /confirm:
    post:
      description: >-
        Confirm some actions for Account managment or Access managment
        (user/create, user/delete, token/get, password/change, login/change). It can return different json's for different situations (see base requests).
      tags:
        - Additional
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/verification_data'
      responses:
        '200':
          description: All is ok
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/token_data'
        '400':
          description: Wrong verification code or user id
    patch:
      description: >-
        Used for resending verification code.
      tags:
        - Additional
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/verification_token'
      responses:
        '200':
          description: All is ok, new code sended and new token created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/verification_token'
        '400':
          description: Wrong verification code or user id
components:
  schemas:
...
    verification_data:
      properties:
        verification_token:
          type: string
        verification_code:
          type: string
      example:
        verification_token: my_verification_token
        verification_code: my_verification_code
    token_data:
      properties:
        token:
          type: string
        refresh_token:
          type: string
      example:
        token: my_token
        refresh_token: my_refresh_token
    verification_token:
      properties:
        verification_token:
          type: string
      example:
        verification_token: my_temporary_token